fixing of the some error in last team user deletion

+++ controllers/team-controllers.js
@@ -462,11 +462,13 @@ const deleteTeamUser = async (req, res) => {
         wantedUser["_id"]
       );
     }
-    //if the deleted user was the last member of the team
+    //if the deleted user was the last member of the team.
+    if(checkedTeam.members.length <= 0){
     //delete all task related to this team
     await TaskModel.deleteMany({ relatedTeam: teamId });
     //delete the empty team
     await TeamModel.findByIdAndDelete(teamId);
+    }
     return res
       .status(200)
       .json({
